fix warnings

+++ src/common/transformations/include/transformations/transpose_sinking/ts_utils.hpp
@@ -66,7 +66,7 @@ namespace sink_forward {
  */
 bool UpdateInputTransposes(const std::shared_ptr<ov::Node>& main_node,
                            const TransposeInputsInfo& transpose_input_info,
-                           std::vector<int> input_indexes = {});
+                           std::vector<size_t> input_indexes = {});
 
 /**
  * @brief Removes @arg input node
@@ -88,7 +88,7 @@ namespace sink_backward {
  */
 ov::NodeVector InsertTransposeBeforeNode(const std::shared_ptr<ov::Node>& main_node,
                                          const std::shared_ptr<ov::opset10::Constant>& transpose_const,
-                                         std::vector<int> input_indexes = {});
+                                         std::vector<size_t> input_indexes = {});
 }  // namespace sink_backward
 
 void UpdateForwardSinkingAbility(const std::shared_ptr<ov::Node>&);
